Imported and rename latest Prometheus library
This commit imports and renames the latest Promtheus library
from the Prometheus Operator github repository. Note that
this version of the Prometheus library has not yet been
released to Charmhub since charmcraft as yet does not
support library name changes. Creating a new library with
a different name is not advised since it is expected that
a specific library will always retain the same library ID.