Grafana dashboard aggregation for machine charms
Read reactive events out of event data, and process multiple
simultaneous requests from the same charm (if they exist)
concurrently.

Also, include the reactive charm dashboard templates in this proxy
charm, and try to apply them if the name matches. Since the old
charm directly queries the prometheus API from the Grafana charm
to fill out optional sections, this is not 1:1 functionality, but
it's a start.

Tweak the algorithm for IDs to remove LZMA headers so they are
actually unique, even when aggregating them all into one data
bucket.

Add interfaces to LMA Proxy for use with Grafana.

Update the documentation to match new prometheus interface names,
add licenses and license headers, update to the newest version
of the prometheus library.

Don't go into blocked state by default. Only if one part of a
relation is missing.

Update the tests to match all of this.