Add grafana to the proxy, update docs/metadata/etc to make it more generic for LMA rather than just Prometheus (#3)
* Update the README to reflect the current state. Add the appropriate interfaces to the charm and metadata

* Add automation boilerplate and a license

* Update the prometheus library

* Grafana dashboard aggregation for machine charms

Read reactive events out of event data, and process multiple
simultaneous requests from the same charm (if they exist)
concurrently.

Also, include the reactive charm dashboard templates in this proxy
charm, and try to apply them if the name matches. Since the old
charm directly queries the prometheus API from the Grafana charm
to fill out optional sections, this is not 1:1 functionality, but
it's a start.

Tweak the algorithm for IDs to remove LZMA headers so they are
actually unique, even when aggregating them all into one data
bucket.

Add interfaces to LMA Proxy for use with Grafana.

Update the documentation to match new prometheus interface names,
add licenses and license headers, update to the newest version
of the prometheus library.

Don't go into blocked state by default. Only if one part of a
relation is missing.

Update the tests to match all of this.

* Doc updates

* Template updates

* Add a different ingress example

* Update .github/workflows/run_tests.yaml

Co-authored-by: Leon <82407168+sed-i@users.noreply.github.com>

* Beef up the tests

* Final library sync

* Cleanup and docs changes

* Update grafana dashboard lib with docs

Co-authored-by: Simon Aronsson <simme@arcticbit.se>
Co-authored-by: Leon <82407168+sed-i@users.noreply.github.com>