Add an NRPE exporter to the Proxy
This involves listening on _four_ more interfaces in order to
successfully get all of the jobs which may be YAML or JSON, and
the command names may either be a string or a dict. It's
documented, but not standardized.

Add a resource for `nrpe_exporter`, and check whether or not the
file is present on the host when a nrpe relation is joined. If it
isn't, then write it to the filesystem and create a systemd unit
to manage it.

`on_stop`, do the reverse -- remove the exporter binary and the
unit.

Adding the scrape config jobs requires some slight changes to
the `MetricsEndpointAggregator` to actually expose methods which
can be used by library users instead of trying to do everything
magically, as additional annotation are required, but duplicating
all of the logic from `prometheus_scrape` inside the NRPE library
does not make any sense versus creating partial structs which
can be exported