Fix error on checking for loki endpoints
If there was an error that HTTPError didn't cover,
such as an error in name resolution,
urllib.request.urlopen may raise URLError instead of HTTPError.
URLError is the parent class for all exceptions that may be raised
by urlopen, so we should catch that exception instead.

Also add logging to help with debugging in the case of such errors.